requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
psycopg2==2.8.3
pycparser==2.19
PySocks==1.7.0
python-dateutil==2.8.1
python-telegram-bot==12.0.0b1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
tornado==6.0.2
urllib3==1.25.3
